Rename feature http -> http-crate
In the context of ureq "http" is a bit too generic feature name to use
for the http crate interop. Instead we use "http-crate" for the 1.0
dependency.
